less movement of code
