+++ core/src/main/scala/org/apache/spark/executor/TaskMetrics.scala
@@ -24,7 +24,6 @@ import org.apache.spark._
 import org.apache.spark.annotation.DeveloperApi
 import org.apache.spark.internal.Logging
 import org.apache.spark.scheduler.AccumulableInfo
-import org.apache.spark.shuffle.TempShuffleReadMetrics
 import org.apache.spark.storage.{BlockId, BlockStatus}
 import org.apache.spark.util._
